fix: bringing changes from detached HEAD commit
A lot of stuff happened, I don't even know when I switched to this
commit to be honest, but I'm glad I noticed it and now it's fixed.
All changes are still here. I made 3 patches with some updates to
two of them, and configured dwm so it looks the same as it did
before the upgrade. Now the clean up process can commence.