fix: rendering artifacts caused by patch mismatch
whilst reapplying patches after st update, I applied changes from a
patch which (of course) didn't have the anysize patch applied.

It caused rendering artifacts in parts of the terminal where anysize
was working. With neovim, it would display a little sliver of my
stausline on the bottom left of the screen. Applying the changes
from anysize into the area where it wasn't applied, it removed that
rendering issue and everything works normally now.

I hope this is the last issue I have to fix with this update.